requirements.txt
alembic==1.18.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
Bootstrap-Flask==2.5.0
cffi==2.0.0
click==8.3.1
dnspython==2.8.0
dominate==2.9.1
email-validator==2.3.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==3.0
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
visitor==0.1.3
Werkzeug==3.1.6
WTForms==3.2.1
